#1089 Move local review providers under tools/local-collab
